backend/requirements.txt
# BrainXR — Python AI Pipeline Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
SimpleITK>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
nibabel>=5.1.0
trimesh>=3.23.0
Pillow>=10.0.0
tqdm>=4.65.0
scipy>=1.11.0
matplotlib>=3.7.0
nilearn>=0.10.0
